+++ src/queryEvents.ts
@@ -164,6 +164,11 @@ export class QueryEvents extends Query<RequestEvents> {
         if (complexQuery instanceof ComplexEventQuery) {
             query.setVal("query", JSON.stringify(complexQuery.getQuery()));
         } else if (_.isString(complexQuery)) {
+            try {
+                JSON.parse(complexQuery);
+            } catch {
+                console.error("Failed to parse the provided string content as a JSON object. Please check the content provided as a parameter to the initWithComplexQuery() method");
+            }
             query.setVal("query", complexQuery);
         } else if (_.isObject(complexQuery)) {
             query.setVal("query", JSON.stringify(complexQuery));